app: move Notify* to app.internal.notify
As we need to access connmgr in almost all methods of the NotifyQueue
class, we pass the connmgr and set it as a class instance attribute.
